% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifLiblineaRL1L2SVC.R
\docType{data}
\name{LearnerClassifLiblineaRL1L2SVC}
\alias{LearnerClassifLiblineaRL1L2SVC}
\alias{mlr_learners_classif.liblinearl1l2svc}
\title{L1-Regularized L2-Loss Support Vector Classification}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifLiblineaRL1L2SVC
}
\description{
A \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for a L1-Regularized L2-Loss Support Vector Classification implemented in \code{\link[LiblineaR:LiblineaR]{LiblineaR::LiblineaR()}} in package \CRANpkg{LiblineaR}.
}
\keyword{datasets}
