% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifLiblineaRL1LogReg.R
\docType{data}
\name{LearnerClassifLiblineaRL1LogReg}
\alias{LearnerClassifLiblineaRL1LogReg}
\alias{mlr_learners_classif.liblinearl1logreg}
\title{L1-Regularized Logistic Regression}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifLiblineaRL1LogReg
}
\description{
A \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for a L1-Regularized Logistic Regression implemented in \code{\link[LiblineaR:LiblineaR]{LiblineaR::LiblineaR()}} in package \CRANpkg{LiblineaR}.
}
\keyword{datasets}
