% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifLiblineaRL2L1SVC.R
\docType{data}
\name{LearnerClassifLiblineaRL2L1SVC}
\alias{LearnerClassifLiblineaRL2L1SVC}
\alias{mlr_learners_classif.liblinearl2l1svc}
\title{L2-Regularized L1-Loss Support Vector Classification}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifLiblineaRL2L1SVC
}
\description{
A \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for a L2-Regularized L1-Loss Support Vector Classification implemented in \code{\link[LiblineaR:LiblineaR]{LiblineaR::LiblineaR()}} in package \CRANpkg{LiblineaR}.
}
\keyword{datasets}
