% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifLiblineaRL2L2SVC.R
\docType{data}
\name{LearnerClassifLiblineaRL2L2SVC}
\alias{LearnerClassifLiblineaRL2L2SVC}
\alias{mlr_learners_classif.liblinearl2l2svc}
\title{L2-Regularized L2-Loss Support Vector Classification}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifLiblineaRL2L2SVC
}
\description{
A \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for a L2-Regularized L2-Loss Support Vector Classification implemented in \code{\link[LiblineaR:LiblineaR]{LiblineaR::LiblineaR()}} in package \CRANpkg{LiblineaR}.
}
\note{
If number of records > number of features \code{type = 2} is faster than \code{type = 1} (Hsu et al. 2016).\cr
If \code{epsilon} is missing and \code{type = 1} (default), \code{epsilon} is set to \code{0.1}. If \code{type = 2}, \code{epsilon} is set to \code{0.01}.
}
\references{
C-W. Hsu, Chang C-C., Lin and C-J (2016)\cr
A Practical Guide to Support Vector Classification\cr
\url{https://www.csie.ntu.edu.tw/~cjlin/papers/guide/guide.pdf}\cr
}
\keyword{datasets}
