% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifLiblineaRL2LogReg.R
\docType{data}
\name{LearnerClassifLiblineaRL2LogReg}
\alias{LearnerClassifLiblineaRL2LogReg}
\alias{mlr_learners_classif.liblinearl2logreg}
\title{L2-Regularized Logistic Regression}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifLiblineaRL2LogReg
}
\description{
A \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for a L2-Regularized Logistic Regression implemented in \code{\link[LiblineaR:LiblineaR]{LiblineaR::LiblineaR()}} in package \CRANpkg{LiblineaR}.
}
\keyword{datasets}
