% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifLiblineaRMultiClassSVC.R
\docType{data}
\name{LearnerClassifLiblineaRMultiClassSVC}
\alias{LearnerClassifLiblineaRMultiClassSVC}
\alias{mlr_learners_classif.liblinearmulticlasssvc}
\title{Support Vector Classification}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifLiblineaRMultiClassSVC
}
\description{
A \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for a Support Vector Classification implemented in \code{\link[LiblineaR:LiblineaR]{LiblineaR::LiblineaR()}} in package \CRANpkg{LiblineaR}.
}
\keyword{datasets}
