% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrLiblineaRL2L1SVR.R
\docType{data}
\name{LearnerRegrLiblineaRL2L1SVR}
\alias{LearnerRegrLiblineaRL2L1SVR}
\alias{mlr_learners_regr.liblinearl2l1svr}
\title{L2-Regularized L1-Loss Support Vector Regression}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrLiblineaRL2L1SVR
}
\description{
A \link[mlr3:LearnerRegr]{mlr3::LearnerRegr} for a L2-Regularized L1-Loss Support Vector Regression implemented in \code{\link[LiblineaR:LiblineaR]{LiblineaR::LiblineaR()}} in package \CRANpkg{LiblineaR}.
}
\keyword{datasets}
