% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrLiblineaRL2L2SVR.R
\docType{data}
\name{LearnerRegrLiblineaRL2L2SVR}
\alias{LearnerRegrLiblineaRL2L2SVR}
\alias{mlr_learners_regr.liblinearl2l2svr}
\title{L2-Regularized L2-Loss Support Vector Regression}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrLiblineaRL2L2SVR
}
\description{
A \link[mlr3:LearnerRegr]{mlr3::LearnerRegr} for a L2-Regularized L2-Loss Support Vector Regression implemented in \code{\link[LiblineaR:LiblineaR]{LiblineaR::LiblineaR()}} in package \CRANpkg{LiblineaR}.
}
\note{
If \code{epsilon} is missing and \code{type = 11} (default), \code{epsilon} is set to \code{0.01}. If \code{type = 12}, \code{epsilon} is set to \code{0.1}.
}
\keyword{datasets}
