% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.liblinear-package}
\alias{mlr3learners.liblinear}
\alias{mlr3learners.liblinear-package}
\title{mlr3learners.liblinear: LiblineaR learners from LiblineaR package for mlr3}
\description{
LiblineaR learners from LiblineaR package for mlr3.
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

}
