% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.liblinear-package}
\alias{mlr3learners.liblinear}
\alias{mlr3learners.liblinear-package}
\title{mlr3learners.liblinear: Learners from LiblineaR package for mlr3}
\description{
Adds `LiblineaR()` from the LiblineaR package to
    mlr3.
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

}
